refactor: 컴포넌트 디렉토리 구조 개선
