--- index.advenced-refactoring.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="ko">
-<head>
-  <meta charset="UTF-8"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-  <title>장바구니로 학습하는 디자인패턴</title>
-  <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body class="bg-gray-100 p-8">
-<div id="root"></div>
-<script type="module" src="/src/refactoring/main.tsx"></script>
-</body>
-</html>
