+++ package.json
@@ -13,10 +13,12 @@
     "test:basic": "vitest src/basic --mode test",
     "test:advanced": "vitest src/advanced --mode test",
     "test:ui": "vitest --ui  --mode test",
-    "build": "tsc -b && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0"
+    "build": "tsc -b && vite build --mode production",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "deploy": "gh-pages -d dist"
   },
   "dependencies": {
+    "gh-pages": "^6.3.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
@@ -24,6 +26,7 @@
     "@testing-library/jest-dom": "^6.6.2",
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
+    "@types/node": "^22.15.0",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.1",
     "@typescript-eslint/eslint-plugin": "^8.10.0",
@@ -43,5 +46,6 @@
     "workerDirectory": [
       "dist"
     ]
-  }
+  },
+  "homepage": "https://ljwoon1211.github.io/front_5th_chapter2-2"
 }
