+++ src/mock-refactoring/components/admin/product/ProductManager.tsx
